Private types and functions

Documentation for CubedSphere.jl's internal interface.

CubedSphere

CubedSphere.find_N — Method

find_N(r; decimals=15)

Return the required number of points we need to consider around the circle of radius r to compute the conformal map series coefficients up to decimals points. The number of points is computed based on the estimate of eq. (B9) in the paper by Rančić et al. [1]. That is, is the smallest integer $N$ (which is chosen to be a power of 2 so that the FFTs are efficient) for which

\[N - \frac{7}{12} \frac{\mathrm{log}_{10}(N)}{\mathrm{log}_{10}(r)} - \frac{r + \mathrm{log}_{10}(A₁ / C)}{-4 \mathrm{log}_{10}(r)} > 0\]

where $r$ is the number of decimals we are aiming for and

\[C = \frac{\sqrt{3} \Gamma(1/3) A₁^{1/3}}{256^{1/3} π}\]

with $A₁$ an estimate of the $Z^1$ Taylor series coefficient of $W(Z)$.

For $A₁ ≈ 1.4771$ we get $C ≈ 0.265$.

References

[1] Rančić et al., Q. J. R. Meteorol., (1996).

CubedSphere.find_taylor_coefficients — Function

find_taylor_coefficients(r = 1 - 1e-7;
                         Niterations = 30,
                         maximum_coefficients = 256,
                         Nevaluations = find_N(r; decimals=15))

Return the Taylor coefficients for the conformal map $Z \to W$ and also of the inverse map, $W \to Z$, where $Z = z^4$ and $W = w^3$. In particular, it returns the coefficients $A_k$ of the Taylor series

\[W(Z) = \sum_{k=1}^\infty A_k Z^k\]

and also coefficients $B_k$ the inverse Taylor series

\[Z(W) = \sum_{k=1}^\infty B_k Z^k\]

The algorithm to obtain the coefficients follows the procedure described in the Appendix of the paper by Rančić et al. [1]

Arguments

r (positional): the radius about the center and the edge of the cube used in the algorithm described by Rančić et al. [1]. r must be less than 1; default: 1 - 10$^{-7}$.

maximum_coefficients (keyword): the truncation for the Taylor series; default: 256.

Niterations (keyword): the number of update iterations we perform on the Taylor coefficients $A_k$; default: 30.

Nevaluations (keyword): the number of function evaluations in over the circle of radius r; default find_N(r; decimals=15); see find_N.

Example

julia> using CubedSphere

julia> using CubedSphere: find_taylor_coefficients

julia> A, B = find_taylor_coefficients(1 - 1e-4);
[ Info: Computing the first 256 coefficients of the Taylor serieses
[ Info: using 32768 function evaluations on a circle with radius 0.9999.
100.0%┣████████████████████████████████████████████┫ 30/30 [00:02<00:00, 12it/s]

julia> A[1:10]
10-element Vector{Float64}:
  1.4771306289227293
 -0.3818351018795475
 -0.05573057838030261
 -0.008958833150428962
 -0.007913155711663374
 -0.004866251689037038
 -0.003292515242976284
 -0.0023548122712604494
 -0.0017587029515141275
 -0.0013568087584722149

Reproducing coefficient table by Rančić et al. (1996)

To reproduce the coefficients tabulated by Rančić et al. [1] use the default values, i.e., $r = 1 - 10^{-7}$.

References

[1] Rančić et al., Q. J. R. Meteorol., (1996).

Elliptic.Jacobi.dn — Method

dn(z::Complex, m::Real)

Compute the Jacobi elliptic function dn(z | m) following Abramowitz & Stegun (1964), Eq. 16.21.4.